KA-511 Rename indexableArticleChunksCreator file
